get_ffmpeg_tool_path moved to ffmpeg utils
